examples/bondai/requirements.txt
bondai
termcolor
python-dotenv
